+++ ActivityMonitor++/GLDataLine.h
@@ -15,7 +15,7 @@
 @property (strong, nonatomic) UIColor *color;
 
 - (id)initWithColor:(UIColor*)aColor forGraph:(GLLineGraph*)aGraph;
-- (void)addLineDataValue:(float)value;
+- (void)addLineDataValue:(double)value;
 - (void)addLineDataArray:(NSArray*)dataArray;
 - (void)resetLineData;
 - (void)render;
